--- core/ui/common/src/main/java/org/mozilla/social/core/ui/common/search/SearchBarTextField.kt
@@ -1,137 +1,0 @@
-package org.mozilla.social.core.ui.common.search
-
-import androidx.compose.foundation.border
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.text.selection.LocalTextSelectionColors
-import androidx.compose.foundation.text.selection.TextSelectionColors
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.LocalTextStyle
-import androidx.compose.material3.TextFieldColors
-import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.takeOrElse
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.unit.dp
-import org.mozilla.social.core.designsystem.theme.MoSoTheme
-
-/**
- * Search bar text field that is essentially the same as the material 3 text field, but
- * with some changes to padding and minimum
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun SearchBarTextField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    readOnly: Boolean = false,
-    textStyle: TextStyle = LocalTextStyle.current,
-    label: @Composable (() -> Unit)? = null,
-    placeholder: @Composable (() -> Unit)? = null,
-    leadingIcon: @Composable (() -> Unit)? = null,
-    trailingIcon: @Composable (() -> Unit)? = null,
-    prefix: @Composable (() -> Unit)? = null,
-    suffix: @Composable (() -> Unit)? = null,
-    supportingText: @Composable (() -> Unit)? = null,
-    isError: Boolean = false,
-    visualTransformation: VisualTransformation = VisualTransformation.None,
-    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
-    keyboardActions: KeyboardActions = KeyboardActions.Default,
-    singleLine: Boolean = false,
-    maxLines: Int = if (singleLine) 1 else Int.MAX_VALUE,
-    minLines: Int = 1,
-    interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
-    shape: Shape = TextFieldDefaults.shape,
-    colors: SearchBarColors = SearchBarTextFieldDefaults.colors(),
-    borderShape: Shape = RoundedCornerShape(90.dp),
-) {
-    // If color is not provided via the text style, use content color as a default
-    val textColor = textStyle.color.takeOrElse {
-        colors.textColor
-    }
-    val mergedTextStyle = textStyle.merge(TextStyle(color = textColor))
-
-    CompositionLocalProvider(LocalTextSelectionColors provides colors.selectionColors) {
-        BasicTextField(
-            value = value,
-            modifier = modifier
-                .border(
-                    width = 1.dp,
-                    color =
-                    if (isError) {
-                        MoSoTheme.colors.borderWarning
-                    } else {
-                        colors.borderColor
-                    },
-                    shape = borderShape,
-                ),
-            onValueChange = onValueChange,
-            enabled = enabled,
-            readOnly = readOnly,
-            textStyle = mergedTextStyle,
-            cursorBrush = SolidColor(colors.cursorColor),
-            visualTransformation = visualTransformation,
-            keyboardOptions = keyboardOptions,
-            keyboardActions = keyboardActions,
-            interactionSource = interactionSource,
-            singleLine = singleLine,
-            maxLines = maxLines,
-            minLines = minLines,
-            decorationBox = @Composable { innerTextField ->
-                // places leading icon, text field with label and placeholder, trailing icon
-                TextFieldDefaults.DecorationBox(
-                    value = value,
-                    visualTransformation = visualTransformation,
-                    innerTextField = innerTextField,
-                    placeholder = placeholder,
-                    label = label,
-                    leadingIcon = leadingIcon,
-                    trailingIcon = trailingIcon,
-                    prefix = prefix,
-                    suffix = suffix,
-                    supportingText = supportingText,
-                    shape = shape,
-                    singleLine = singleLine,
-                    enabled = enabled,
-                    isError = isError,
-                    interactionSource = interactionSource,
-                    colors = colors.inputFieldColors,
-                    contentPadding = PaddingValues(0.dp)
-                )
-            }
-        )
-    }
-}
-
-internal object SearchBarTextFieldDefaults {
-    @Composable
-    fun colors(): SearchBarColors =
-        SearchBarColors(
-            cursorColor = MoSoTheme.colors.textPrimary,
-            textColor = MoSoTheme.colors.textPrimary,
-            selectionColors = LocalTextSelectionColors.current,
-            borderColor = MoSoTheme.colors.borderInputEnabled,
-            inputFieldColors = MoSoSearchBarDefaults.colors().inputFieldColors,
-        )
-}
-
-data class SearchBarColors(
-    val cursorColor: Color,
-    val textColor: Color,
-    val selectionColors: TextSelectionColors,
-    val borderColor: Color,
-    val inputFieldColors: TextFieldColors,
-)
